- Adds code to mock getting a json component representation from a remote service
