+++ desktopApp/src/jvmMain/kotlin/com.macaosoftware.sdui.app/Main.kt
@@ -26,53 +26,32 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.singleWindowApplication
 import com.macaosoftware.component.DesktopComponentRender
-import com.macaosoftware.component.navbar.BottomNavigationComponent
-import com.macaosoftware.component.navbar.BottomNavigationComponentDefaults
 import com.macaosoftware.platform.DesktopBridge
+import com.macaosoftware.sdui.app.data.SduiRemoteService
 import com.macaosoftware.sdui.app.di.SharedKoinContext
-import com.macaosoftware.sdui.app.viewmodel.factory.AppBottomNavigationViewModelFactory
 import com.pablichj.incubator.amadeus.Database
 import com.pablichj.incubator.amadeus.storage.DriverFactory
 import com.pablichj.incubator.amadeus.storage.createDatabase
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
-import kotlinx.serialization.json.JsonObject
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.addJsonObject
-import kotlinx.serialization.json.buildJsonObject
-import kotlinx.serialization.json.putJsonArray
-import org.koin.core.context.startKoin
-import org.koin.dsl.koinApplication
 import org.koin.dsl.module
 import kotlin.system.exitProcess
 
 fun main() {
 
-    val rootComponentJson = SduiService().getRootJson()
+    val rootComponentJson = SduiRemoteService.getRootJson()
+    val rootComponent = SduiLocalService().getComponentInstanceOf(rootComponentJson)
 
     val windowState = WindowState(size = DpSize(500.dp, 800.dp))
+    val desktopBridge = DesktopBridge()
 
     val database = runBlocking { createDatabase(DriverFactory()) }
-
-    val storageModule = module { single <Database> { database } }
+    val storageModule = module { single<Database> { database } }
     /*startKoin {
         modules(storageModule)
     }*/
     SharedKoinContext.initKoin(
         listOf(storageModule)
     )
-
-    val navBarComponent = BottomNavigationComponent(
-        viewModelFactory = AppBottomNavigationViewModelFactory(
-            sduiHandler = AppBottomSduiHandler(rootComponentJson),
-            database = database,
-            bottomNavigationStatePresenter = BottomNavigationComponentDefaults.createBottomNavigationStatePresenter(
-                dispatcher = Dispatchers.Main
-            )
-        ),
-        content = BottomNavigationComponentDefaults.BottomNavigationComponentView
-    )
-    val desktopBridge = DesktopBridge()
 
     singleWindowApplication(
         title = "Amadeus Desktop Demo",
@@ -135,7 +114,7 @@ fun main() {
                     }
                 }
                 DesktopComponentRender(
-                    rootComponent = navBarComponent,
+                    rootComponent = rootComponent,
                     windowState = windowState,
                     desktopBridge = desktopBridge
                 )
